example/requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
langchain-community
chromadb
python-dotenv
SQLAlchemy
pymysql
huggingface-hub

